Added the jobs for restaurants_apis
