+++ app/controllers/itineraries_controller.rb
@@ -95,6 +95,8 @@ class ItinerariesController < ApplicationController
                                 day:)
         if category.save!
           set_restaurant(food_time)
+          AccommodationApiJob.perform_later(food_time, itineraries_params, max_price_generator, @itinerary, category) # <- The job is queued
+
         else
           # Category Failed
         end
@@ -156,93 +158,10 @@ class ItinerariesController < ApplicationController
     end
   end
 
-  def set_yelp_price(price_string)
-    case
-    when price_string.nil? || price_string == " " then return 0
-    when price_string == "￥" then return 10
-    when price_string == "￥￥" then return 30
-    when price_string == "￥￥￥" then return 60
-    when price_string == "￥￥￥￥" then return 100
-    end
-  end
-
-  def check_restaurant_api_location(api_location, location)
-    case
-    when api_location.nil? then return location
-    when !api_location.nil? then return api_location
-    end
-  end
+  
 
   def set_restaurant(food_time)
-    if food_time == "Lunch"
-      restaurant_budget = max_price_generator / 10
-      set_restaurant_budget = []
 
-      if restaurant_budget >= 60
-        set_restaurant_budget = "1, 2, 3, 4"
-      elsif restaurant_budget >= 30 && restaurant_budget < 60
-        set_restaurant_budget = "1, 2, 3"
-      elsif restaurant_budget >= 10 && restaurant_budget < 30
-        set_restaurant_budget = "1, 2"
-      else
-        set_restaurant_budget = "1"
-      end
-      restaurants = RestaurantApiService.new(location: @itinerary.location,
-                                             keyword: "Best Lunch restaurants",
-                                             price: set_restaurant_budget)
-
-      begin
-        restaurants_results = restaurants.call
-        restaurants_selected = restaurants_results.sample
-      rescue
-        retry
-      end
-
-      restaurants_selected["location"]["display_address"].nil? ? restaurant_location = location : restaurant_location = restaurants_selected["location"]["display_address"].first
-
-      Content.create!(name: restaurants_selected["name"],
-                      price: set_yelp_price(restaurants_selected["price"]),
-                      location: restaurant_location,
-                      rating: restaurants_selected["rating"],
-                      category: Category.last,
-                      description: restaurants_selected["categories"].first["title"],
-                      api: "",
-                      status: 0)
-
-    else
-      restaurant_budget = max_price_generator / 5
-      set_restaurant_budget = []
-
-      if restaurant_budget >= 60
-        set_restaurant_budget = "1, 2, 3, 4"
-      elsif restaurant_budget >= 30 && restaurant_budget < 60
-        set_restaurant_budget = "1, 2, 3"
-      elsif restaurant_budget >= 10 && restaurant_budget < 30
-        set_restaurant_budget = "1, 2"
-      else
-        set_restaurant_budget = "1"
-      end
-      restaurants = RestaurantApiService.new(location: @itinerary.location,
-                                             keyword: "Best Dinner restaurants",
-                                             price: set_restaurant_budget)
-      begin
-        restaurants_results = restaurants.call
-        restaurants_selected = restaurants_results.sample
-      rescue
-        retry
-      end
-
-      restaurants_selected["location"]["display_address"].nil? ? restaurant_location = location : restaurant_location = restaurants_selected["location"]["display_address"].first
-
-      Content.create!(name: restaurants_selected["name"],
-                      price: set_yelp_price(restaurants_selected["price"]),
-                      location: restaurant_location,
-                      rating: restaurants_selected["rating"],
-                      category: Category.last,
-                      description: restaurants_selected["categories"].first["title"],
-                      api: "",
-                      status: 0)
-    end
   end
 
   def set_activity
