+++ app/jobs/activity_api_job.rb
@@ -1,0 +1,7 @@
+class ActivityApiJob < ApplicationJob
+  queue_as :default
+
+  def perform(*args)
+    # Do something later
+  end
+end
