+++ test/jobs/restaurant_api_job_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class RestaurantApiJobTest < ActiveJob::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
